Generator

Documentation

sampleMaze :: Maze

geraAleatorios :: Int -> Int -> [Int]

Given a seed returns a list of n integer randomly generated

nrAleatorio :: Int -> Int

Given a seed returns an integer randomly generated

subLista :: Int -> [a] -> [[a]]

Converts a list into a list of list of size n

convertePeca :: Int -> Piece

Converts an integer number into a Peca

printCorridor :: Corridor -> String

Converts a Corredor to a string

converteCorredor :: [Int] -> Corridor

Converts a list of integers into a Corredor

converteLabirinto :: [[Int]] -> Maze

Converts a list of lists of integers into a Labirinto

geraLabirinto :: Int -> Int -> Int -> Maze

Generates a random maze

imprimeLabirinto :: Maze -> IO ()

prints a maze