Tarefa1

Documentation

generateMaze :: Int -> Int -> Int -> Maze

Gera o Labirinto através da largura e da altura pretendida e ainda de uma seed

constroiCasa :: Int -> Maze

Dado uma largura do labirinto dá a respetiva casa do fantasma

geraCorredor :: Int -> Corridor

Gera um corredor preenchido por walls da largura pretendida

geraParedes :: Maze -> Int -> Int -> Maze

A um labirinto coloca Paredes com o respetivo túnel

colocaTunel :: Int -> Int -> Piece

Verifica se está a meio da altura do labirinto para verificar se é necessário colocar Wall ou Empty

labPar :: Int -> Int -> Maze -> Maze

Verifica se a Largura do labirinto e par para realizar a substituição da respetiva casa dos fantasmas

colocaCasa

Arguments

:: Int
Largura do labirinto

-> Int
Altura do labirinto

-> Maze
Labirinto a ser substituido

-> Int
Variável para determinar quando é necessário substituir o corredor

-> Int
Variável para determinar que lista e necessária ir usar da função ConstroiCasa

-> Maze

Obtém o meio do labirinto onde é necessário alterar para a casa dos fantasmas quando o labirinto é par

colocaCasa'

Arguments

:: Int
Largura do labirinto

-> Int
Altura do labirinto

-> Maze
Labirinto a ser substituido

-> Int
Variável para determinar quando é necessário substituir o corredor

-> Int
Variável para determinar que lista e necessária ir usar da função ConstroiCasa

-> Maze

Obtém o meio do labirinto onde é necessário alterar para a casa dos fantasmas quando o labirinto é ímpar

subCasa

Arguments

:: Int
Largura do labirinto

-> Int
Nº da lista que se vai buscar à função constroiCasa

-> Corridor
Corredor que se pretende substituir

-> Corridor

Obtém a lista da casa dos fantasmas necessária para substituir no maze

testar :: Int -> Int -> Int -> IO ()

testes :: (Eq t, Num t) => Int -> Int -> Int -> t -> [Char]

teste1 :: Foldable t => [t a] -> Int -> Int -> [Char]

teste2 :: [[Piece]] -> Int -> p -> [Char]

teste3 :: [[Piece]] -> p -> Int -> [Char]