Tarefa2

Documentation

play :: Play -> State -> State

Efetua uma jogada movendo o Pacman espicificado segundo uma orientação

movePlayer

Arguments

:: State

-> Player
Player que vai ser atualizado

-> Orientation
Orientação do movimento

-> State
Devolve o State atualizado

Verifica se o Pacman tem a mesma orientação da jogada e se pode transitar na direção desejada

newPlayerState

Arguments

:: [Player]
Lista inicial de Players

-> Player
É necessário o Player que queremos atualizar para atulizar o seu próprio estado na lista original

-> Piece
Usamos a peça para onde o Pacman se vai mover para realizar a ação pretendida

-> Coords
Coordenadas do meio do labirinto para caso um fantasma seja comido reapareça na casa dos fantasmas

-> [Player]

Realiza os efeitos no estado do jogador segundo o que contém a proxima peça

correctMaze

Arguments

:: Maze

-> Coords
Coordenadas de onde o Pacman parte inicialmente

-> Maze

Corrige o labirinto quando o Pacman transita de peça, colocando Empty de onde ele sai

typePlace

Arguments

:: Maze

-> Coords
Coordenadas originais do Pacman

-> Orientation
Orientação da jogada

-> Piece

Obtem a peça para onde o Pacman se vai mover

nextPlace

Arguments

:: Int
largura do maze

-> Coords
Coordenadas originais do Pacman

-> Orientation
Orientação da jogada

-> Coords

Obtem as coordenadas para onde o Pacmana se vai mover

updateGhost

Arguments

:: [Player]
Lista com o estado do Pacman atualizado segundo a peça

-> Player
É necessário para informações e comparações

-> [Player]
Usamos uma lista vazia para manter a ordenação da lista original

-> Coords
Coordenadas do meio do labirinto caso um fantasma seja comido

-> Int
Variável para dizer quantos fantasmas comeu ou se o Pacman foi comido

-> [Player]

Atualiza todos os estados dos Ghosts para Dead, quando o Pacman come comida grande e verifica também se existe algum Ghost na mesma coordenada do Pacman

aux :: [Player] -> Player -> [Player]

É uma função auxiliar para as funções updateGhost e gHost que corrige o estado do Pacman quando embate com um fantasma Alive

gHost

Arguments

:: [Player]
Lista com o estado do Pacman atualizado segundo a peça

-> Player
É necessário para informações e comparações

-> Coords
Coordenadas do meio do labirinto caso o fantasma seja comido

-> [Player]
Usamos uma lista vazia para manter a ordenação da lista original

-> Int
Variável para dizer quantos fantasmas comeu ou se o Pacman foi comido

-> [Player]

Verifica e determina o efieto da existência e do estado de um fantasma na mesma coordenada do Pacman

mazetest :: Maze

statest :: State

testar :: (Eq a, Num a) => a -> State

testes :: (Eq a, Num a) => State -> a -> State